+++ src/ai/backend/client/cli/run.py
@@ -846,23 +846,23 @@ def start(image, session_id, owner,                                 # base args
 
 
 @main.command(aliases=['rm', 'kill'])
-@click.argument('sess_id_or_alias', metavar='SESSID', nargs=-1)
+@click.argument('session_ids', metavar='SESSID', nargs=-1)
 @click.option('-o', '--owner', '--owner-access-key', metavar='ACCESS_KEY',
               help='Specify the owner of the target session explicitly.')
 @click.option('-s', '--stats', is_flag=True,
               help='Show resource usage statistics after termination')
-def terminate(sess_id_or_alias, owner, stats):
+def terminate(session_ids, owner, stats):
     '''
     Terminate the given session.
 
-    SESSID: session ID or its alias given when creating the session.
+    SESSID: session ID given/generated when creating the session.
     '''
     print_wait('Terminating the session(s)...')
     with Session() as session:
         has_failure = False
-        for sess in sess_id_or_alias:
+        for session_id in session_ids:
             try:
-                kernel = session.Kernel(sess, owner)
+                kernel = session.Kernel(session_id, owner)
                 ret = kernel.destroy()
             except BackendAPIError as e:
                 print_error(e)
@@ -874,16 +874,49 @@ def terminate(sess_id_or_alias, owner, stats):
             except Exception as e:
                 print_error(e)
                 has_failure = True
-            if has_failure:
-                sys.exit(1)
         else:
-            print_done('Done.')
+            if not has_failure:
+                print_done('Done.')
             if stats:
                 stats = ret.get('stats', None) if ret else None
                 if stats:
                     print(_format_stats(stats))
                 else:
                     print('Statistics is not available.')
+        if has_failure:
+            sys.exit(1)
+
+
+@main.command()
+@click.argument('session_ids', metavar='SESSID', nargs=-1)
+def restart(session_ids):
+    '''
+    Restart the given session.
+
+    SESSID: session ID given/generated when creating the session.
+    '''
+    print_wait('Restarting the session(s)...')
+    with Session() as session:
+        has_failure = False
+        for sess in session_ids:
+            try:
+                kernel = session.Kernel(sess)
+                kernel.restart()
+            except BackendAPIError as e:
+                print_error(e)
+                if e.status == 404:
+                    print_info(
+                        'If you are an admin, use "-o" / "--owner" option '
+                        'to terminate other user\'s session.')
+                has_failure = True
+            except Exception as e:
+                print_error(e)
+                has_failure = True
+        else:
+            if not has_failure:
+                print_done('Done.')
+        if has_failure:
+            sys.exit(1)
 
 
 @main.command()